Change num_flops to int64 Allows for num_flops to be calculated for larger matrices

+++ src/benchmark_base.f90
@@ -1,8 +1,9 @@
 module benchmark_base
+    use iso_fortran_env, only: int64
     implicit none
 
     type, abstract :: Benchmark
-        integer :: num_flops
+        integer(int64) :: num_flops
         character(len=16) :: name
     contains
         procedure(setup_interface), deferred :: setup
